requirements.txt
google-generativeai
duckduckgo-search
beautifulsoup4
requests
python-dotenv
